LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/mathml/presentation-markup/mrow/mrow-fallback-expected.txt


PASS Invalid <mfrac> should lay out as an mrow (count == 0)
FAIL Invalid <mfrac> should lay out as an mrow (count == 1) assert_approx_equals: inline size expected 10 +/- 1 but got 0
FAIL Invalid <mfrac> should lay out as an mrow (count == 3) assert_approx_equals: inline size expected 60 +/- 1 but got 0
FAIL Invalid <mfrac> should lay out as an mrow (count == 4) assert_approx_equals: inline size expected 100 +/- 1 but got 0
FAIL Invalid <mfrac> should lay out as an mrow (count == 5) assert_approx_equals: inline size expected 150 +/- 1 but got 0
PASS Invalid <mroot> should lay out as an mrow (count == 0)
FAIL Invalid <mroot> should lay out as an mrow (count == 1) assert_approx_equals: inline size expected 10 +/- 1 but got 0
FAIL Invalid <mroot> should lay out as an mrow (count == 3) assert_approx_equals: inline size expected 60 +/- 1 but got 0
FAIL Invalid <mroot> should lay out as an mrow (count == 4) assert_approx_equals: inline size expected 100 +/- 1 but got 0
FAIL Invalid <mroot> should lay out as an mrow (count == 5) assert_approx_equals: inline size expected 150 +/- 1 but got 0
PASS Invalid <munder> should lay out as an mrow (count == 0)
FAIL Invalid <munder> should lay out as an mrow (count == 1) assert_approx_equals: inline size expected 10 +/- 1 but got 0
FAIL Invalid <munder> should lay out as an mrow (count == 3) assert_approx_equals: inline size expected 60 +/- 1 but got 0
FAIL Invalid <munder> should lay out as an mrow (count == 4) assert_approx_equals: inline size expected 100 +/- 1 but got 0
FAIL Invalid <munder> should lay out as an mrow (count == 5) assert_approx_equals: inline size expected 150 +/- 1 but got 0
PASS Invalid <mover> should lay out as an mrow (count == 0)
FAIL Invalid <mover> should lay out as an mrow (count == 1) assert_approx_equals: inline size expected 10 +/- 1 but got 0
FAIL Invalid <mover> should lay out as an mrow (count == 3) assert_approx_equals: inline size expected 60 +/- 1 but got 0
FAIL Invalid <mover> should lay out as an mrow (count == 4) assert_approx_equals: inline size expected 100 +/- 1 but got 0
FAIL Invalid <mover> should lay out as an mrow (count == 5) assert_approx_equals: inline size expected 150 +/- 1 but got 0
PASS Invalid <munderover> should lay out as an mrow (count == 0)
FAIL Invalid <munderover> should lay out as an mrow (count == 1) assert_approx_equals: inline size expected 10 +/- 1 but got 0
FAIL Invalid <munderover> should lay out as an mrow (count == 2) assert_approx_equals: inline size expected 30 +/- 1 but got 0
FAIL Invalid <munderover> should lay out as an mrow (count == 4) assert_approx_equals: inline size expected 100 +/- 1 but got 0
FAIL Invalid <munderover> should lay out as an mrow (count == 5) assert_approx_equals: inline size expected 150 +/- 1 but got 0
PASS Invalid <msub> should lay out as an mrow (count == 0)
FAIL Invalid <msub> should lay out as an mrow (count == 1) assert_approx_equals: inline size expected 10 +/- 1 but got 0
FAIL Invalid <msub> should lay out as an mrow (count == 3) assert_approx_equals: inline size expected 60 +/- 1 but got 0
FAIL Invalid <msub> should lay out as an mrow (count == 4) assert_approx_equals: inline size expected 100 +/- 1 but got 0
FAIL Invalid <msub> should lay out as an mrow (count == 5) assert_approx_equals: inline size expected 150 +/- 1 but got 0
PASS Invalid <msup> should lay out as an mrow (count == 0)
FAIL Invalid <msup> should lay out as an mrow (count == 1) assert_approx_equals: inline size expected 10 +/- 1 but got 0
FAIL Invalid <msup> should lay out as an mrow (count == 3) assert_approx_equals: inline size expected 60 +/- 1 but got 0
FAIL Invalid <msup> should lay out as an mrow (count == 4) assert_approx_equals: inline size expected 100 +/- 1 but got 0
FAIL Invalid <msup> should lay out as an mrow (count == 5) assert_approx_equals: inline size expected 150 +/- 1 but got 0
PASS Invalid <msubsup> should lay out as an mrow (count == 0)
FAIL Invalid <msubsup> should lay out as an mrow (count == 1) assert_approx_equals: inline size expected 10 +/- 1 but got 0
FAIL Invalid <msubsup> should lay out as an mrow (count == 2) assert_approx_equals: inline size expected 30 +/- 1 but got 0
FAIL Invalid <msubsup> should lay out as an mrow (count == 4) assert_approx_equals: inline size expected 100 +/- 1 but got 0
FAIL Invalid <msubsup> should lay out as an mrow (count == 5) assert_approx_equals: inline size expected 150 +/- 1 but got 0
PASS Invalid <mmultiscripts> should lay out as an mrow (count == 0)
FAIL Invalid <mmultiscripts> should lay out as an mrow (count == 2) assert_approx_equals: inline size expected 30 +/- 1 but got 0
FAIL Invalid <mmultiscripts> should lay out as an mrow (count == 4) assert_approx_equals: inline size expected 100 +/- 1 but got 0
FAIL Invalid <mmultiscripts> should lay out as an mrow (count == 7) assert_approx_equals: inline size expected 220 +/- 1 but got 0
FAIL Invalid <mmultiscripts> should lay out as an mrow (count == 9) assert_approx_equals: inline size expected 390 +/- 1 but got 0
FAIL Invalid mmultiscripts should lay out as an mrow (first in-flow child is an <mprescripts>) assert_approx_equals: inline size expected 100 +/- 1 but got 66.375
FAIL Invalid mmultiscripts should lay out as an mrow (one of the even number of children after the first <mprescripts> is an <mprescripts>) assert_approx_equals: inline size expected 220 +/- 1 but got 0
